requirements.txt
mysql-connector-python==8.0.22
protobuf==3.4.0